+++ idea/testData/quickfix/createFromUsage/createTypeParameter/inReferencedDeclaration/missingArguments.kt
@@ -1,0 +1,4 @@
+// "Create type parameter in class 'X'" "false"
+// ERROR: 2 type arguments expected for class X<T, U> defined in root package
+class X<T, U>
+fun Y(x: X<<caret>String>) {}